requirements.txt
flask
werkzeug
requests
flask-mysql
mysql-connector-python


